feat: change type of Paging.after

+++ src/main/kotlin/com/kroffle/knitting/infra/persistence/design/repository/R2dbcDesignRepository.kt
@@ -28,7 +28,7 @@ class R2dbcDesignRepository(private val repository: DBDesignRepository) : Design
                     repository
                         .findAllByKnitterIdAndIdBefore(
                             knitterId = knitterId,
-                            id = paging.after,
+                            id = paging.after.toLong(),
                             pageable = pageRequest,
                         )
                 } else {
